fix both of anton's bugs: * make([100]int) was being compiled to 	make([]int), kind of. * []this = [100]that was working for any this, that.
turned up a typo in pipe_test.go

R=ken
OCL=27081
CL=27081

+++ src/lib/io/pipe_test.go
@@ -148,7 +148,7 @@ func testPipeReadClose(t *testing.T, async bool) {
 	} else {
 		delayClose(t, w, c);
 	}
-	var buf [64]int;
+	var buf [64]byte;
 	n, err := r.Read(buf);
 	<-c;
 	if err != nil {
